Allow configuration of java executable
The path of the java executable can now be configured in the ProtobufExtension and/or specific GenerateProtoTask instances.

* GenerateProtoTask gains the javaExecutablePath Property,
* ProtobufExtension gains the javaExecutablePath Property and the defaultJavaExecutablePath provider, which provides the default path using the same logic as previous versions
* computeJavaExePath moved from GenerateProtoTask to ProtobufExtension since it is now only used in ProtobufExtension
* isWindows moved from GenerateProtoTask to Util since it is now used in GenerateProtoTask and ProtobufExtension
